# Set functions now accept variadic arguments