## Summary

<!-- Describe your changes in detail here, if it closes an open issue, include "Closes #<issue>" -->
Changed `Difference`, `Intersection` and `Union` in the `set` package to now accept variadic arguments meaning they work with any number of input sets.

Also tweaked them to claw back (and in most cases overall improve) performance to account for now looping over input sets. The only exception is `Intersection` which is now a tiny bit slower, but the trade off IMO is worth it

```plaintext
❯ benchstat before.txt after.txt
goos: darwin
goarch: amd64
pkg: github.com/FollowTheProcess/collections/set
cpu: Intel(R) Core(TM) i5-8257U CPU @ 1.40GHz
               │ before.txt  │              after.txt              │
               │   sec/op    │   sec/op     vs base                │
Intersection-8   69.55µ ± 0%   75.25µ ± 1%   +8.21% (p=0.000 n=30)
Insert-8         73.80µ ± 1%   72.46µ ± 0%   -1.81% (p=0.000 n=30)
Union-8          135.0µ ± 1%   116.3µ ± 1%  -13.89% (p=0.000 n=30)
Difference-8     68.35µ ± 1%   61.81µ ± 1%   -9.57% (p=0.000 n=30)
geomean          82.96µ        79.12µ        -4.63%

               │  before.txt  │              after.txt              │
               │     B/op     │     B/op      vs base               │
Intersection-8   23.46Ki ± 0%   23.46Ki ± 0%       ~ (p=0.892 n=30)
Insert-8         46.62Ki ± 0%   46.62Ki ± 0%       ~ (p=0.432 n=30)
Union-8          49.05Ki ± 0%   48.92Ki ± 0%  -0.28% (p=0.000 n=30)
Difference-8     23.46Ki ± 0%   23.33Ki ± 0%  -0.56% (p=0.000 n=30)
geomean          33.49Ki        33.42Ki       -0.21%

               │ before.txt │              after.txt              │
               │ allocs/op  │ allocs/op   vs base                 │
Intersection-8   40.00 ± 0%   40.00 ± 0%       ~ (p=1.000 n=30) ¹
Insert-8         65.00 ± 0%   65.00 ± 0%       ~ (p=1.000 n=30) ¹
Union-8          91.00 ± 0%   88.00 ± 0%  -3.30% (p=0.000 n=30)
Difference-8     40.00 ± 0%   37.00 ± 0%  -7.50% (p=0.000 n=30)
geomean          55.46        53.94       -2.75%
¹ all samples are equal
```